# Use `memchr` to optimize `get_text_slice`