--- a/fluent-syntax/src/parser/macros.rs
+++ b/fluent-syntax/src/parser/macros.rs
@@ -9,3 +9,9 @@ macro_rules! get_current_byte {
         $s.source.as_ref().as_bytes().get($s.ptr)
     };
 }
+
+macro_rules! get_remaining_bytes {
+    ($s:expr) => {
+        $s.source.as_ref().as_bytes().get($s.ptr..)
+    };
+}