@clubby789 2024-02-21
Some notable improvements [when used in `rustc`](https://github.com/rust-lang/rust/pull/121351):

|                                    | mean  | range          | count |
|:----------------------------------:|:-----:|:--------------:|:-----:|
| Regressions ❌ <br /> (primary)    | -     | -              | 0     |
| Regressions ❌ <br /> (secondary)  | -     | -              | 0     |
| Improvements ✅ <br /> (primary)   | -0.3% | [-0.3%, -0.2%] | 3     |
| Improvements ✅ <br /> (secondary) | -1.5% | [-4.1%, -0.3%] | 41    |
| All ❌✅ (primary)                 | -0.3% | [-0.3%, -0.2%] | 3     |